third_party/WebKit/LayoutTests/fast/dom/custom/document-register-basic-expected.txt

Testing document.register() basic behaviors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof fooConstructor is 'function'
PASS fooConstructor.prototype.__proto__ is HTMLElement.prototype
PASS fooConstructor.prototype.thisIsPrototype is true
PASS document.register('x-bad-a', HTMLElement) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS document.register('x-bad-b', fooConstructor) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS document.register('x-bad-c', { prototype: proto }) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS document.register('x-bad-d', { prototype: {} }) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS document.register('x-bad-e', { prototype: Object.create(Document.prototype) }) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS fooConstructor() threw exception TypeError: DOM object constructor cannot be called as a function..
PASS createdFoo.__proto__ is fooConstructor.prototype
PASS createdFoo.constructor is fooConstructor
PASS createdFoo.tagName is 'X-FOO'
PASS createdFoo.textContent is 'Hello'
PASS createdFoo.lastChild is childDiv
PASS parsedFoo.__proto__ is fooConstructor.prototype
PASS parsedFoo.tagName is 'X-FOO'
PASS parsedFoo.someProperty is container.firstChild.someProperty
PASS barConstructor !== fooConstructor is true
PASS createdBar.tagName is 'X-BAR'
PASS createdBaz.tagName is 'X-BAZ'
PASS createdBaz.thisIsPrototype is true
PASS createdBaz.thisIsAlsoPrototype is true
PASS createdUpperBar.constructor is barConstructor
PASS createdUpperBar.tagName is 'X-BAR'
PASS createdMixedBar.constructor is barConstructor
PASS createdMixedBar.tagName is 'X-BAR'
PASS container.firstChild.constructor is barConstructor
PASS container.firstChild.tagName is 'X-BAR'
PASS container.lastChild.constructor is barConstructor
PASS container.lastChild.tagName is 'X-BAR'
PASS (new fooConstructor).tagName is 'X-FOO'
PASS (new barConstructor).tagName is 'X-BAR'
PASS (new bazConstructor).tagName is 'X-BAZ'
PASS successfullyParsed is true

TEST COMPLETE
